+++ src/main/java/tech/aroma/application/service/reactions/actions/RunThroughInboxAction.java
@@ -21,8 +21,8 @@ import org.apache.thrift.TException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import sir.wellington.alchemy.collections.lists.Lists;
-import tech.aroma.application.service.operations.reactions.MessageReactor;
-import tech.aroma.application.service.reactions.MatchAlgorithm;
+import tech.aroma.application.service.reactions.MessageReactor;
+import tech.aroma.application.service.reactions.matchers.MatchAlgorithm;
 import tech.aroma.data.ReactionRepository;
 import tech.aroma.thrift.Message;
 import tech.aroma.thrift.User;
