+++ src/main/java/tech/aroma/application/service/reactions/matchers/MatchAlgorithm.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package tech.aroma.application.service.reactions;
+package tech.aroma.application.service.reactions.matchers;
 
 import java.util.List;
 import java.util.Objects;
